+++ NEWS.md
@@ -1,4 +1,4 @@
-- added cc_and() and cc_or() as useful wrappers
+- added cc_and(), cc_or(), co_and() and co_or() wrappers
 - set domain = NA for messages, warnings and errors
 
 # err 0.0.1